fix: rename FilterAuthor and FilterTag namespaces to improve organization
