+++ src/Http/Actions/Author/AuthorGetListAction.php
@@ -7,7 +7,7 @@ use CSlant\Blog\Api\Http\Requests\Author\AuthorGetListRequest;
 use CSlant\Blog\Api\Http\Resources\Author\ListAuthorResource;
 use CSlant\Blog\Api\OpenApi\Schemas\Resources\Author\ListAuthorResourceSchema;
 use CSlant\Blog\Api\Services\AuthorService;
-use CSlant\Blog\Api\Supports\FilterAuthor;
+use CSlant\Blog\Api\Supports\Filters\FilterAuthor;
 use CSlant\Blog\Core\Http\Actions\Action;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
